modify something

+++ app/Http/Controllers/TasksController.php
@@ -127,9 +127,10 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-        $task=Task::find($id);
-        $task->delete();
+        $task=\App\Micropost::find($id);
         
-        return redirect('/');
+        if(\Auth::id() === $task->user_id){
+            $task->delete();
+        }
     }
 }
